kth smallest element

+++ ARRAY/3-minMax.cpp
@@ -2,7 +2,8 @@
 using namespace std;
 int getMax(int arr[], int n)
 {
-   int maxi = INT_MIN;
+   int maxi = INT64_MIN;
+
    for(int i=0; i<n; i++)
    {
      maxi = max(maxi, arr[i]);
@@ -11,7 +12,7 @@ int getMax(int arr[], int n)
 }
 int getMin(int arr[], int n)
 {
-  int mini = INT_MAX;
+  int mini = INT64_MAX;
   for(int i=0; i<n; i++)
   {
     mini = min(mini, arr[i]);
